Add files via upload
arial10x10.png is a file in the libtcod folder that I had to add to the RL folder on my computer. Odds are I could have just added the file path in the engine.py program where it's called, but I wanted to keep things as close as possible to the tutorial.