Add files via upload
There was something wrong with my GameMap class. For some reason it wouldn't make any of the rooms, so when I ran engine.py it just had a full screen of unmovable tiles. Eventually I just cut and pasted the code form the devs. github and then it worked. No idea why